Use STS client for GetAccountId
Fixes #139. The `GetAccountIdE` method was originally using `iamClient.GetUser` to try to get the current AWS account, but this method doesn't work if you're authenticated via an IAM Role or something that isn't an IAM User. This PR updates the method to use `sts.GetCallerIdentity`, which should work no matter how you're authenticated.