Fix backend config params
Terraform seems to require that backend config arguments are passed using the syntax `-backend-config=key=value`. That is, with an equals sign as a separator. The syntax we use elsewhere with a space as a separator (e.g., `-var key=value`) doesn't seem to work for backend config. Added example + test to verify too.