Switch retryable errors map back to use string as key
Golang can't natively serialize a regular expression... So when we try to save a `packer.Options` or `terraform.Options` object with regex in the retryable errors map, we get an error. So, we switch back to using strings as keys, and have the `DoRetryWithRetryableErrors` methdo compile those strings into regular expressions.