Use ReadLine to get stable output scanning
This commit switches from bufio/Scanner to bufio/Reader to read command
execution's stdout and stderr. This removes the need for a `MaxLineSize`
in the buffer.

The reason why this was replaced is because I was able to make a
command's execution hang forever. I'm not quite sure what caused it, but
it manifested in the following output:

```
TestModule/FreshInstallSubtest 2020-04-14T08:41:37Z command.go:87: Running command terraform with args [show -no-color -json /tmp/620357425/p.tfplan]
```

The call never returned, and when I had a look at the code I decided to
remove the Scanner and replace it with a Reader instead. Though it may
be a bit slower (https://stackoverflow.com/a/48598005), it fixes the bug
and makes the API easier to use, as no MaxLineSize needs to be passed
through anymore.

The output in question is coming from a `terraform show -json` call,
which resulted in an output that was one line and ~500KB. I saved that
json-output to a file and tried reproducing the bug with a cut-down
version of the shell-module (basically using os.Open and using that to
create the scanner), but it works fine like that.
I am able to reproduce the bug by copying the plan to my local machine
(the output above is from a gitlab runner) using the .go file here:
https://gist.github.com/tommyknows/0e0e995c03f75ac474aa135af1870e53

Sadly, I cannot share the terraform plan output. The terraform plan is
~50KB, the json version is ~500KB, but because the terraform workspace
needs to be initialised and it's a plan from internal infrastructure,
I cannot share it.

Signed-off-by: Ramon Rüttimann <ramon@nine.ch>