fix for handling -no-color and -var options
`terraform apply` expects `plan file` to be presented as the last
argument, currently -no-color option comes after the plan file. this
commit fixes that.

when `terraform apply` is provided with a plan file, we should not pass
`-var` and `-var-file` options anymore. it is a violation. this commit
fixes that.