feat: add AssertTgPlanAllExitCode
this adds an assertion to use with TgPlanAllE.

since run-all plan with `-detailed-exitcode` can return a `0` or `2` on successful plans,
this assertions checks for either acceptable value and asserts if true or false
depending on the assertion under test, designated by the `assertTrue` input arg.

this has been passed around as a helper function for simple regression testing using plan all
across many modules, seemed like time to share it with the community.